+++ Gemfile
@@ -1,4 +1,4 @@
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in interactive_replacer.gemspec
+# Specify your gem's dependencies in irep.gemspec
 gemspec
