--- lib/interactive_replacer.rb
@@ -1,8 +1,0 @@
-require "interactive_replacer/version"
-require "interactive_replacer/cli"
-require "interactive_replacer/search"
-require "interactive_replacer/replace"
-
-module InteractiveReplacer
-  # Your code goes here...
-end
