+++ lib/irep.rb
@@ -1,0 +1,8 @@
+require "irep/version"
+require "irep/cli"
+require "irep/search"
+require "irep/replace"
+
+module InteractiveReplacer
+  # Your code goes here...
+end
