+++ lib/irep/cli.rb
@@ -1,7 +1,7 @@
 require 'optparse'
-require 'interactive_replacer/search'
-require 'interactive_replacer/replace'
-require 'interactive_replacer/interface'
+require 'irep/search'
+require 'irep/replace'
+require 'irep/interface'
 
 module InteractiveReplacer
   class CLI
