+++ test/test_irep.rb
